% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_total_effect.r
\name{estimate_all_total_effects}
\alias{estimate_all_total_effects}
\title{全変数間の総合因果効果を一括推定}
\usage{
estimate_all_total_effects(X, lingam_result)
}
\arguments{
\item{X}{元データ (n_samples x n_features)}

\item{lingam_result}{direct_lingam() の返り値}
}
\value{
総合因果効果の行列 (行: 結果変数, 列: 原因変数)
}
\description{
全変数間の総合因果効果を一括推定
}
\examples{
# ここの実行例を書き直し
pk <- make_prior_knowledge(6, exogenous_variables = c(4))

# ここの実行例を書き直し
pk <- make_prior_knowledge(6,
  exogenous_variables = "x3",
  sink_variables = c("x1", "x4"),
  paths = list(c("x3", "x0"), c("x3", "x2")),
  no_paths = list(c("x5", "x2")),
  labels = c("x0", "x1", "x2", "x3", "x4", "x5")
)
}
